Move tests to a single file
